rustdoc: qualify generated `Result`
This helps when using a custom `--test-builder` to fetch and
then insert the source code elsewhere, in case there is
a custom `Result` being used.

For instance, in the kernel, we currently generate tests like:

```rust
pub fn rust_kernel_doctest_workqueue_rs_120_0() {
    // ...

    // Many tests need the prelude, so provide it by default.
    #[allow(unused)]
    use kernel::prelude::*;

    {
        // ... contains the output of `rustdoc` ...
        fn main() {
            #[allow(non_snake_case)]
            fn _doctest_main_rust_kernel_workqueue_rs_120_0()
                -> Result<(), impl core::fmt::Debug> {
                // ...
            }
            _doctest_main_rust_kernel_workqueue_rs_120_0().unwrap()
        }
        main();
    }
}
```

In the `kernel` prelude, we happen to have a custom `Result` in
the prelude, thus we need to manually tweak the source code to
qualify the path.

One could go further and use a global path, possibly taking into
account Edition 2015 as needed, but given `Debug` is already using
the canonical path, just do the same for `Result` too.

In the future, if `rustdoc` provided a fine-grained way of
extracting the test contents, this may not be needed. For the
moment, however, it should not hurt.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>